add requestSubscription
